+++ sway-lsp/src/lib.rs
@@ -7,6 +7,8 @@ pub mod config;
 mod core;
 pub mod error;
 mod server;
+#[cfg(test)]
+mod tests;
 mod traverse;
 pub mod utils;
 use server::Backend;
